+++ tests/Unit/Config/CheckoutGatewayTest.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace Payavel\Checkout\Tests\Unit;
+namespace Payavel\Checkout\Tests\Unit\Config;
 
+use Payavel\Checkout\Tests\Unit\TestCheckoutGateway;
 use Payavel\Orchestration\Tests\Traits\CreatesConfigServiceables;
 use Payavel\Orchestration\Tests\Traits\SetsConfigDriver;
 
-class ConfigCheckoutGatewayTest extends TestCheckoutGateway
+class CheckoutGatewayTest extends TestCheckoutGateway
 {
     use CreatesConfigServiceables;
     use SetsConfigDriver;
