+++ tests/Unit/Database/CheckoutGatewayTest.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace Payavel\Checkout\Tests\Unit;
+namespace Payavel\Checkout\Tests\Unit\Database;
 
+use Payavel\Checkout\Tests\Unit\TestCheckoutGateway;
 use Payavel\Orchestration\Tests\Traits\CreatesDatabaseServiceables;
 use Payavel\Orchestration\Tests\Traits\SetsDatabaseDriver;
 
-class DatabaseCheckoutGatewayTest extends TestCheckoutGateway
+class CheckoutGatewayTest extends TestCheckoutGateway
 {
     use CreatesDatabaseServiceables;
     use SetsDatabaseDriver;
